requirements.txt
Django
django-filter
djangorestframework
coreapi
